requirements.txt
torch
torchvision
tqdm
Pillow
scikit-learn
numpy
pandas
huggingface_hub
streamlit
dotenv